Update .gitignore
First you will be asked to select a directory which would be used to write or read files to that folder a new folder inside the selected folder is created with Name Darwin and time stamp of the time.You need to enter column name of the Target . Then a list of all the variables pop up double click only Categorical Variables and close the popup(target column name is modified to target you need to select that too). You will be asked to select row or column which delete's the row's or column's which have more than 30% of NA's. Rest is taken care by the tool. When execution is done Open the directory that was selected by you in start. It would have Folder named Darwin in which you will have Layer-6 folder with output csv file where all the metrics with combinations used are displayed. 
If any doubts comment below